DAL: Ajout service Jeux + interface + mapper + SP_GetAll

+++ DB_LudoShare/Jeux.sql
@@ -3,11 +3,11 @@
 	[Jeux_Id] UNIQUEIDENTIFIER NOT NULL DEFAULT NEWID() PRIMARY KEY,
 	[Nom] NVARCHAR(64) NOT NULL,
 	[Description] NVARCHAR(512) NOT NULL,
-	[AgeMin] TINYINT CHECK (AgeMin >= 0),
-	[AgeMax] TINYINT CHECK (AgeMax >= AgeMin),
-	[NbJoueurMin] TINYINT CHECK (NbJoueurMin >= 0),
-	[NbJoueurMax] TINYINT CHECK (NbJoueurMax >= NbJoueurMin),
-	[DureeMinute] SMALLINT CHECK (DureeMinute > 0),
+	[AgeMin] INT CHECK (AgeMin >= 0),
+	[AgeMax] INT CHECK (AgeMax >= AgeMin),
+	[NbJoueurMin] INT CHECK (NbJoueurMin >= 0),
+	[NbJoueurMax] INT CHECK (NbJoueurMax >= NbJoueurMin),
+	[DureeMinute] INT CHECK (DureeMinute > 0),
 	[DateCreation] DATETIME2 NOT NULL DEFAULT GETDATE(),
 	CONSTRAINT CK_AgeMax CHECK (AgeMax >= AgeMin) 
 )
